+++ backend/src/app/db/sql/0011-mod-contact-table.sql
@@ -1,5 +1,3 @@
 ALTER TABLE contact ADD COLUMN validated_at timestamptz NULL;
 
 UPDATE contact SET validated_at=now();
-
-ALTER TABLE contact ALTER COLUMN validated_at SET NOT NULL;