+++ backend/src/app/db/sql/0013-mod-contact-table-add-email-index.sql
@@ -1,0 +1,3 @@
+CREATE UNIQUE INDEX contact__params_email__uidx
+    ON contact (owner_id, (params->>'~:email'))
+ WHERE type = 'email';