+++ backend/src/app/util/emails.clj
@@ -14,27 +14,174 @@
    [cuerdas.core :as str]
    [app.common.spec :as us]
    [app.common.exceptions :as ex]
-   [app.util.template :as tmpl]))
+   [app.util.template :as tmpl])
+  (:import
+   java.util.Properties
+   javax.mail.Message
+   javax.mail.Transport
+   javax.mail.Message$RecipientType
+   javax.mail.PasswordAuthentication
+   javax.mail.Session
+   javax.mail.internet.InternetAddress
+   javax.mail.internet.MimeMultipart
+   javax.mail.internet.MimeBodyPart
+   javax.mail.internet.MimeMessage))
 
-;; --- Impl.
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+;; Email Building
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+
+(defn build-address
+  [v charset]
+  (try
+    (cond
+      (string? v)
+      (InternetAddress. v nil charset)
+
+      (map? v)
+      (InternetAddress. (:addr v)
+                        (:name v)
+                        (:charset v charset))
+
+      :else
+      (throw (ex-info "Invalid address" {:data v})))
+    (catch Exception e
+      (throw (ex-info "Invalid address" {:data v} e)))))
+
+(defn- resolve-recipient-type
+  [type]
+  (case type
+    :to  Message$RecipientType/TO
+    :cc  Message$RecipientType/CC
+    :bcc Message$RecipientType/BCC))
+
+(defn- assign-recipient
+  [^MimeMessage mmsg type address charset]
+  (if (sequential? address)
+    (reduce #(assign-recipient %1 type %2 charset) mmsg address)
+    (let [address (build-address address charset)
+          type    (resolve-recipient-type type)]
+      (.addRecipient mmsg type address)
+      mmsg)))
+
+(defn- assign-recipients
+  [mmsg {:keys [to cc bcc charset] :or {charset "utf-8"} :as params}]
+  (cond-> mmsg
+    (some? to)  (assign-recipient :to to charset)
+    (some? cc)  (assign-recipient :cc cc charset)
+    (some? bcc) (assign-recipient :bcc bcc charset)))
+
+(defn- assign-from
+  [mmsg {:keys [from charset] :or {charset "utf-8"}}]
+  (when from
+    (let [from (build-address from charset)]
+      (.setFrom ^MimeMessage mmsg ^InternetAddress from))))
+
+(defn- assign-reply-to
+  [mmsg {:keys [defaut-reply-to]} {:keys [reply-to charset] :or {charset "utf-8"}}]
+  (let [reply-to (or reply-to defaut-reply-to)]
+    (when reply-to
+      (let [reply-to (build-address reply-to charset)
+            reply-to (into-array InternetAddress [reply-to])]
+        (.setReplyTo ^MimeMessage mmsg reply-to)))))
+
+(defn- assign-subject
+  [mmsg {:keys [subject charset] :or {charset "utf-8"}}]
+  (assert (string? subject) "subject is mandatory")
+  (.setSubject ^MimeMessage mmsg
+               ^String subject
+               ^String charset))
+
+(defn- assign-extra-headers
+  [^MimeMessage mmsg {:keys [headers custom-data] :as params}]
+  (let [headers (assoc headers "X-Sereno-Custom-Data" custom-data)]
+    (reduce-kv (fn [^MimeMessage mmsg k v]
+                 (doto mmsg
+                   (.addHeader (name k) (str v))))
+               mmsg
+               headers)))
+
+(defn- assign-body
+  [^MimeMessage mmsg {:keys [body charset] :or {charset "utf-8"}}]
+  (let [mpart (MimeMultipart. "mixed")]
+    (cond
+      (string? body)
+      (let [bpart (MimeBodyPart.)]
+        (.setContent bpart ^String body (str "text/plain; charset=" charset))
+        (.addBodyPart mpart bpart))
+
+      (map? body)
+      (let [bpart (MimeBodyPart.)]
+        (.setContent bpart
+                     ^String (:content body)
+                     ^String (str (:type body "text/plain") "; charset=" charset))
+        (.addBodyPart mpart bpart))
+
+      :else
+      (throw (ex-info "Unsupported type" {:body body})))
+    (.setContent mmsg mpart)
+    mmsg))
+
+(defn- build-message
+  [cfg session params]
+  (let [mmsg (MimeMessage. ^Session session)]
+    (assign-recipients mmsg params)
+    (assign-from mmsg params)
+    (assign-reply-to mmsg cfg params)
+    (assign-subject mmsg params)
+    (assign-extra-headers mmsg params)
+    (assign-body mmsg params)
+    (.saveChanges mmsg)
+    mmsg))
+
+(defn- opts->props
+  [{:keys [username tls host port timeout default-from]
+    :or {timeout 30000}
+    :as opts}]
+  (reduce-kv
+   (fn [^Properties props k v]
+     (if (nil? v)
+       props
+       (doto props (.put ^String k  ^String (str v)))))
+   (Properties.)
+   {"mail.smtp.auth" (boolean username)
+    "mail.smtp.starttls.enable" tls
+    "mail.smtp.host" host
+    "mail.smtp.port" port
+    "mail.smtp.from" default-from
+    "mail.smtp.user" username
+    "mail.user" username
+    "mail.host" host
+    "mail.smtp.timeout" timeout
+    "mail.smtp.connectiontimeout" timeout}))
+
+(defn smtp-session
+  [{:keys [debug] :or {debug false} :as opts}]
+  (let [props   (opts->props opts)
+        session (Session/getInstance props)]
+    (.setDebug session debug)
+    session))
+
+(defn smtp-message
+  [cfg message]
+  (let [^Session session (smtp-session cfg)]
+    (build-message cfg session message)))
+
+;; TODO: specs for smtp config
+
+(defn send!
+  [cfg message]
+  (let [^MimeMessage message (smtp-message cfg message)]
+    (Transport/send message (:username cfg) (:password cfg))
+    nil))
+
+;; ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+;; Template Email Building
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
 
 (def ^:private email-path "emails/%(id)s/%(lang)s.%(type)s")
 
-(defn- build-base-email
-  [data context]
-  (when-not (s/valid? ::parsed-email data)
-    (ex/raise :type :internal
-              :code :template-parse-error
-              :hint "Seems like the email template has invalid data."
-              :contex data))
-  {:subject (:subject data)
-   :content (cond-> []
-              (:body-text data) (conj {:type "text/plain"
-                                       :value (:body-text data)})
-              (:body-html data) (conj {:type "text/html"
-                                       :value (:body-html data)}))})
-
-(defn- render-email-part
+(defn- render-email-template-part
   [type id context]
   (let [lang (:lang context :en)
         path (str/format email-path {:id (name id)
@@ -43,20 +190,14 @@
     (some-> (io/resource path)
             (tmpl/render context))))
 
-(defn- impl-build-email
+(defn- build-email-template
   [id context]
   (let [lang (:lang context :en)
-        subj (render-email-part :subj id context)
-        html (render-email-part :html id context)
-        text (render-email-part :txt id context)]
+        subj (render-email-template-part :subj id context)
+        text (render-email-template-part :txt id context)]
     {:subject subj
-     :content (cond-> []
-                text (conj {:type "text/plain"
-                            :value text})
-                html (conj {:type "text/html"
-                            :value html}))}))
-
-;; --- Public API
+     :body {:type "text/plain"
+            :content text}}))
 
 (s/def ::priority #{:high :low})
 (s/def ::to (s/or :sigle ::us/email
@@ -64,13 +205,14 @@
 (s/def ::from ::us/email)
 (s/def ::reply-to ::us/email)
 (s/def ::lang string?)
+(s/def ::custom-data ::us/string)
 
 (s/def ::context
   (s/keys :req-un [::to]
-          :opt-un [::reply-to ::from ::lang ::priority]))
+          :opt-un [::reply-to ::from ::lang ::priority ::custom-data]))
 
-(defn build
-  ([id] (build id {}))
+(defn template-factory
+  ([id] (template-factory id {}))
   ([id extra-context]
    (s/assert keyword? id)
    (fn [context]
@@ -82,18 +224,21 @@
                             (extra-context)
                             extra-context)
                           context)
-           email (impl-build-email id context)]
+           email   (build-email-template id context)]
        (when-not email
          (ex/raise :type :internal
                    :code :email-template-does-not-exists
                    :hint "seems like the template is wrong or does not exists."
-                   ::id id))
+                   :context {:id id}))
        (cond-> (assoc email :id (name id))
-         (:track-id context)
-         (assoc :track-id (:track-id context))
+         (:custom-data context)
+         (assoc :custom-data (:custom-data context))
 
-         (string? (:to context))
-         (assoc :to [(:to context)])
+         (:from context)
+         (assoc :from (:from context))
 
-         (vector? (:to context))
+         (:reply-to context)
+         (assoc :reply-to (:reply-to context))
+
+         (:to context)
          (assoc :to (:to context)))))))