+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 - Add contact verification process.
 - Add single contact unsubscribe link to notification emails.
 - Add complete unsubscribe link to notification emails.
+- Fix many issues in contact creation.
+- Prevent duplicate contact creation.
+- Rework internal email sending.
+- Integrate with AWS SES bounce/complaint notifications.
 
 
 ## v1-2020.10.07-0